requirements.txt
transformers==4.36.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.6.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
datasets==2.16.1
openai>=1.30.0
python-dotenv==1.0.0
tqdm==4.66.1
evaluate==0.4.1
matplotlib==3.8.2
seaborn==0.13.0
langchain>=0.1.0
langchain-openai>=0.1.0
streamlit>=1.28.0 